Move code from debug-mode.php to 0-load.php
